KEL-123 moved DataBaseTestUtils to test folder
